feat(tarko): refactor event processor architecture
